requirements.txt
azure-functions>=1.18.0,<2.0.0
azure-storage-blob>=12.19.0,<13.0.0
azure-cognitiveservices-vision-computervision>=0.9.0,<1.0.0
azure-cognitiveservices-vision-customvision>=3.1.0,<4.0.0
openai>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
Pillow>=10.0.0,<13.0.0
msrest>=0.7.0,<1.0.0
pytest
ruff
pylint